backend/CREZE/requirements.txt
asgiref==3.8.1
boto3==1.35.2
botocore==1.35.2
click==8.1.7
Django==5.0
django-auditlog==3.0.0
django-cors-headers==4.3.1
django-database-url==1.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
